# internal/httpserve/route/openapi.yaml
openapi: 3.1.0
info:
  title: Datum OpenAPI 3.1.0 Specifications
  description: Documentation for Datum's API services
  termsOfService: https://datum.net/terms
  contact:
    name: Datum Support
    url: https://datum.net/support
    email: support@datum.net
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1
servers:
  - url: https://api.datum.net/v1
paths:
...
  /register:
    post:
      operationId: register
      tags:
        - Register
      security:
        - {}
      summary: Register User
      description: Register a new user and their personal organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first-name
                - last-name
                - email
                - password
              properties:
                first-name:
                  type: string
                last-name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User and organization created successfully. The details of the new user are returned
          content:
            application/json:
              schema:
                type: object
                required:
                  - user_id
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          description: A user with the email address already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorReply"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /refresh:
    post:
      operationId: refresh
      tags:
        - Authentication
      security:
        - {}
      summary: Refresh access token
      description: >
        Access tokens have an expiration of an hour, in order to maintain long running sessions, a refresh token that is tied with a specific access token is issued at the same time. The refresh token may be used to obtain a new access/refresh token pair, thereby extending the user or api key session without requiring credentials to be authenticated a second time.

        Note that the claims that are returned during refresh may not be the same claims as the original access token (e.g. if the role of a user has changed during the session). This means that an operation that was previously allowed by the original access token may not be allowed after refresh.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: a JWT token that is signed by Datum and linked to an access token. Has a "Not Before" and "Expires" timestamp that bound when the refresh token can be used
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginReply"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          description: Could not authenticate refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorReply"
        "500":
          $ref: "#/components/responses/InternalError"
  /forgot-password:
    post:
      operationId: forgot-password
      tags:
        - ForgotPassword
      security:
        - {}
      summary: Request a password reset
      description: Ability for a user to trigger a password re-set flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        "204":
          description: Password reset received; no content is returned
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    ErrorReply:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: False if there is an error, true otherwise
        error:
          type: string
          description: The error message with more detail about why the request failed
    LoginReply:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          description: A JWT token signed by Datum that authenticates the user and contains claims that authorize them
        refresh_token:
          type: string
          description: A JWT token signed by Datum that can be used to obtain a new access/refresh token pair
  responses:
...
    InvalidRequest:
      description: Invalid request, user should repair the request and try again
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorReply"
    InternalError:
      description: Internal error, admins need to address the issue on the server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorReply"